Refactor cleaner useSession export path
