+++ apps/admin-ui/src/hooks/index.ts
@@ -11,3 +11,4 @@ export { useFilteredReservationList } from "./useFilteredReservationList";
 export { useStaffReservationMutation } from "./useStaffReservationMutation";
 export { useUnitResources } from "./useUnitResources";
 export { useGetFilterSearchParams } from "./useGetFilterSearchParams";
+export { useSession } from "./useSession";
