+++ apps/ui/components/LoginFragment.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useTranslation } from "next-i18next";
 import { signIn } from "common/src/browserHelpers";
-import { useSession } from "@/hooks/auth";
+import { useSession } from "@/hooks";
 import { Button, ButtonSize } from "hds-react";
 import { getLocalizationLang } from "common/src/helpers";
