+++ apps/ui/components/common/Navigation.tsx
@@ -11,7 +11,7 @@ import {
 } from "hds-react";
 import { useTranslation } from "next-i18next";
 import styled from "styled-components";
-import { useSession } from "@/hooks/auth";
+import { useSession } from "@/hooks";
 import { type CurrentUserQuery } from "@gql/gql-types";
 import Logo from "common/src/components/Logo";
 import { useRouter } from "next/router";
