+++ apps/ui/hooks/index.ts
@@ -8,3 +8,4 @@ export { useCurrentUser } from "./useCurrentUser";
 export { useAvailableTimes } from "./useAvailableTimes";
 export { useRemoveStoredReservation } from "./useRemoveStoredReservation";
 export { useBlockingReservations } from "./useBlockingReservations";
+export { useSession } from "./useSession";
