+++ apps/ui/pages/reservations/index.tsx
@@ -13,7 +13,7 @@ import {
   ReservationTypeChoice,
 } from "@gql/gql-types";
 import { filterNonNullable } from "common/src/helpers";
-import { useSession } from "@/hooks/auth";
+import { useSession } from "@/hooks";
 import { ReservationCard } from "@/components/reservation";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
 import { toApiDate } from "common/src/common/util";
